RAG-Chatbot/RAG-Chatbot/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-core
sentence-transformers
faiss-cpu
pypdf
python-dotenv
groq